requirements.txt
Django==4.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
python-decouple==3.6
python-dotenv==0.21.0
gunicorn==20.1.0
django-heroku==0.3.1